Corrected `minmax` and simplified caching.
Using both `player` and `status` to decide on MIN or MAX was causing
problems with the algorithm. Player `1` should always be maximizing and
player `-1` should always be minimizing. This simplifies the caching,
and in addition the board state determines whose turn it is removing the
need for multiple dictionaries. The boards are now stored as strings to
avoid any potential mutibility problems.